requirements.txt
addict==2.2.1
albumentations==0.4.6
appdirs==1.4.4
asynctest==0.13.0
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
cityscapesScripts==1.5.0
codecov==2.1.9
coverage==5.2.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
flake8==3.8.3
idna==2.10
imagecorruptions==1.1.0
imageio==2.9.0
imgaug==0.4.0
iniconfig==1.0.1
isort==5.4.2
kiwisolver==1.2.0
kwarray==0.5.9
matplotlib==3.3.1
mccabe==0.6.1
mmcv==0.5.1
-e git+https://github.com/KoryakovDmitry/mmdetection.git@2edeecaaccf474d143deabf8fd50fcf6160a2ab9#egg=mmdet
more-itertools==8.4.0
networkx==2.5
numpy==1.19.1
opencv-python==4.4.0.42
ordered-set==4.0.2
packaging==20.4
pandas==1.1.1
Pillow==6.2.1
pluggy==0.13.1
py==1.9.0
pycocotools==2.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytesseract==0.3.5
pytest==6.0.1
pytest-cov==2.10.1
pytest-runner==5.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
scikit-image==0.17.2
scipy==1.5.2
Shapely==1.7.1
six==1.15.0
terminaltables==3.1.0
tifffile==2020.8.13
toml==0.10.1
torch==1.4.0
torchvision==0.5.0
ubelt==0.9.1
urllib3==1.25.10
xdoctest==0.13.0
yapf==0.30.0
